Replace uppergateOut, uppergateIn, lowergateOut, lowergateIn by upperLayerOut, upperLayerIn, lowerLayerOut, lowerLayerIn. This criteria has been used in inet-framework integration branch

+++ src/linklayer/ieee80211/radio/Ieee80211NewRadioModel.cc
@@ -35,7 +35,7 @@ Ieee80211NewRadioModel::~Ieee80211NewRadioModel()
 
 void Ieee80211NewRadioModel::initializeFrom(cModule *radioModule)
 {
-    snirThreshold = dB2fraction(radioModule->par("snirThreshold"));
+    snirThreshold = dB2fraction(radioModule->par("snirThreshold").doubleValue());
 
 
     if (strcmp("SHORT", radioModule->par("WifiPreambleMode").stringValue())==0)
